# docs/swagger.yaml
definitions:
...
  model.Role:
    properties:
      alias:
        description: 别名
        type: string
      createdAt:
        description: 创建时间
        type: integer
      createdBy:
        description: 创建人
        type: string
      createdId:
        description: 创建人ID
        type: string
      id:
        description: 主键ID
        type: string
      name:
        description: 名称
        type: string
      remark:
        description: 备注
        type: string
      sort:
        description: 排序
        type: integer
      status:
        description: 状态 1、正常 2、禁用
        type: integer
      updatedAt:
        description: 更新时间
        type: integer
      updatedBy:
        description: 更新人
        type: string
    type: object
...
  model.User:
    properties:
      avatar:
        description: 用户头像
        type: string
      createdAt:
        description: 创建时间
        type: integer
      createdBy:
        description: 创建人
        type: string
      createdId:
        description: 创建人ID
        type: string
      email:
        description: 邮箱
        type: string
      emailVerified:
        description: 邮箱是否验证 1、否 2、是
        type: integer
      id:
        description: 主键ID
        type: string
      jobNum:
        description: 工号
        type: string
      loginTime:
        description: 最后登录时间
        type: integer
      nickName:
        description: 昵称
        type: string
      phone:
        description: 手机号
        type: string
      phoneVerified:
        description: 手机号是否验证 1、否 2、是
        type: integer
      pwdTime:
        description: 修改密码时间
        type: integer
      realName:
        description: 真实名称
        type: string
      sex:
        description: 性别 1、男 2、女
        type: integer
      status:
        description: 状态 1、正常 2、禁用
        type: integer
      updatedAt:
        description: 更新时间
        type: integer
      updatedBy:
        description: 更新人
        type: string
      username:
        description: 账号
        type: string
    type: object
...
  request.AssignRolesParams:
    properties:
      roleIds:
        description: 角色ID列表
        items:
          type: string
        type: array
      userIds:
        description: 用户ID列表
        items:
          type: string
        type: array
    required:
    - roleIds
    - userIds
    type: object
...
  request.IdsReq:
    properties:
      ids:
        description: ID数组
        items:
          type: string
        type: array
    type: object
...
  request.ResetPasswordParams:
    properties:
      ids:
        description: 父ID
        items:
          type: string
        type: array
      password:
        description: 密码
        type: string
    required:
    - ids
    - password
    type: object
...
  request.RolePageParams:
    properties:
      alias:
        description: 别名
        type: string
      name:
        description: 名称
        type: string
      page:
        description: 页码
        type: integer
      pageSize:
        description: 每页大小
        type: integer
      searchCount:
        description: 查询总数 0，是 1，否
        type: integer
      status:
        description: 状态 1、正常 2、禁用
        type: integer
    type: object
  request.RoleParams:
    properties:
      name:
        description: 名称
        type: string
    type: object
  request.RoleResourceSetParams:
    properties:
      alias:
        description: 别名
        type: string
      id:
        description: 主键ID
        type: string
      name:
        description: 名称
        type: string
      remark:
        description: 备注
        type: string
      resourceIds:
        description: 菜单权限ID集合
        items:
          type: integer
        type: array
      sort:
        description: 排序
        type: integer
      status:
        description: 状态 1、正常 2、禁用
        type: integer
    required:
    - alias
    - name
    type: object
...
  request.UserAssignSetParams:
    properties:
      avatar:
        description: 用户头像
        type: string
      departmentIds:
        description: 部门ID集合
        items:
          type: integer
        type: array
      email:
        description: 邮箱
        type: string
      id:
        description: 主键ID
        type: string
      jobNum:
        description: 工号
        type: string
      nickName:
        description: 昵称
        type: string
      password:
        description: 用户登录密码
        type: string
      phone:
        description: 手机号
        type: string
      realName:
        description: 真实名称
        type: string
      roleIds:
        description: 角色ID集合
        items:
          type: integer
        type: array
      sex:
        description: 性别 1、男 2、女
        type: integer
      username:
        description: 用户登录名
        type: string
    required:
    - username
    type: object
  request.UserPageParams:
    properties:
      departmentId:
        description: 部门ID
        type: string
      email:
        description: 邮箱
        type: string
      jobNum:
        description: 工号
        type: string
      nickName:
        description: 昵称
        type: string
      page:
        description: 页码
        type: integer
      pageSize:
        description: 每页大小
        type: integer
      phone:
        description: 手机号
        type: string
      realName:
        description: 真实名称
        type: string
      searchCount:
        description: 查询总数 0，是 1，否
        type: integer
      sex:
        description: 性别 1、男 2、女
        type: integer
      status:
        description: 状态 1、正常 2、禁用
        type: integer
      username:
        description: 账号
        type: string
    type: object
  request.UserParams:
    properties:
      status:
        example: "0"
        type: string
      username:
        type: string
    type: object
info:
  contact: {}
  description: artificial intelligence message push service
  title: apipig api
  version: 1.0.0
paths:
...
  /v1/sys/role/delete:
    post:
      consumes:
      - application/json
      parameters:
      - description: ID集合
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.IdsReq'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"删除成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 删除角色
      tags:
      - Role
  /v1/sys/role/get:
    get:
      consumes:
      - application/json
      parameters:
      - description: 角色ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 根据id获取角色
      tags:
      - Role
  /v1/sys/role/list:
    post:
      consumes:
      - application/json
      parameters:
      - description: 角色列表查询参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.RoleParams'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 获取角色列表
      tags:
      - Role
  /v1/sys/role/page:
    post:
      consumes:
      - application/json
      parameters:
      - description: 页码, 每页大小等分页参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.RolePageParams'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 分页获取角色列表
      tags:
      - Role
  /v1/sys/role/resource-set:
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建角色
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.RoleResourceSetParams'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 角色创建修改并菜单权限分配
      tags:
      - Role
  /v1/sys/role/update:
    post:
      consumes:
      - application/json
      parameters:
      - description: 角色信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.Role'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"设置成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 更新角色
      tags:
      - Role
...
  /v1/sys/user/assign-roles:
    post:
      consumes:
      - application/json
      parameters:
      - description: 分配角色
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.AssignRolesParams'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"设置成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 分配角色
      tags:
      - User
  /v1/sys/user/assign-set:
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户分配设置参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.UserAssignSetParams'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 用户分配设置
      tags:
      - User
  /v1/sys/user/delete:
    post:
      consumes:
      - application/json
      parameters:
      - description: ID集合
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.IdsReq'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"删除成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 删除用户
      tags:
      - User
  /v1/sys/user/get:
    get:
      consumes:
      - application/json
      parameters:
      - description: 用户ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 根据id获取用户
      tags:
      - User
  /v1/sys/user/info:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 获取当前用户信息
      tags:
      - User
  /v1/sys/user/list:
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户列表查询参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.UserParams'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 获取用户列表
      tags:
      - User
  /v1/sys/user/page:
    post:
      consumes:
      - application/json
      parameters:
      - description: 页码, 每页大小等分页参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.UserPageParams'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 分页获取用户列表
      tags:
      - User
  /v1/sys/user/reset-password:
    post:
      consumes:
      - application/json
      parameters:
      - description: 密码重置
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.ResetPasswordParams'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"设置成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 密码重置
      tags:
      - User
  /v1/sys/user/update:
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.User'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"设置成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 更新用户
      tags:
      - User
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: accessToken
    type: apiKey
swagger: "2.0"